Add calendar style
