--- client/src/utils/timeDisplayHelpers.js
@@ -1,4 +1,0 @@
-export const createTimeSlectors = () => {
-    const times = []
-    
-}
